comments addressed
